@@ -24,8 +24,9 @@
     "@babel/runtime": "^7.17.7",
     "@mna/content-server": "^2.3.5",
     "babel-preset-es2015": "^6.24.1",
-    "react": "^16.6.0",
+    "react": "^18.1.0",
     "react-collapsed": "^3.3.1",
-    "react-dom": "^16.6.0"
+    "react-dom": "^18.1.0",
+    "react-jsx-parser": "^1.29.0"
   }
 }
